Refactor and isolate FT donation flow
