+++ src/features/donation/constants.ts
@@ -1,3 +1,12 @@
+import {
+  type DonationAllocationStrategy,
+  DonationAllocationStrategyEnum,
+  type DonationAllocationStrategyOption,
+  type DonationGroupAllocationStrategy,
+  DonationGroupAllocationStrategyEnum,
+  type DonationGroupAllocationStrategyOption,
+} from "./types";
+
 export const DONATION_MIN_NEAR_AMOUNT = 0.1;
 
 export const DONATION_INSUFFICIENT_BALANCE_ERROR =
@@ -8,3 +17,36 @@ export const DONATION_MIN_NEAR_AMOUNT_ERROR = `The minimum donation amount is ${
 export const DONATION_MAX_MESSAGE_LENGTH = 100;
 
 export const DONATION_BASE_STORAGE_DEPOSIT_FLOAT = 0.012;
+
+export const DONATION_ALLOCATION_STRATEGIES: Record<
+  DonationAllocationStrategy,
+  DonationAllocationStrategyOption
+> = {
+  full: {
+    label: "Direct donation",
+    value: DonationAllocationStrategyEnum.full,
+  },
+
+  share: {
+    label: "Quadratically matched donation",
+    hintIfDisabled: "(no pots available)",
+    value: DonationAllocationStrategyEnum.share,
+  },
+};
+
+export const DONATION_GROUP_ALLOCATION_STRATEGIES: Record<
+  DonationGroupAllocationStrategy,
+  DonationGroupAllocationStrategyOption
+> = {
+  evenly: {
+    label: "Evenly",
+    hint: "(Allocate funds evenly across multiple projects)",
+    value: DonationGroupAllocationStrategyEnum.evenly,
+  },
+
+  manually: {
+    label: "Manually",
+    hint: "(Specify amount for each project)",
+    value: DonationGroupAllocationStrategyEnum.manually,
+  },
+};
